hswidget: move hscommon/hsutil/custom_widget.go into hswidget
